POST '/auth/email-confirmation' initial endpoint setup.
