+++ api_server/auth/cruds/__init__.py
@@ -1,0 +1,5 @@
+from auth.cruds.email_confirmation_tokens import EmailConfirmationTokenCRUD
+
+__all__ = [
+    'EmailConfirmationTokenCRUD',
+]
